+++ R/require_namespace.r
@@ -1,1 +1,0 @@
-requireNamespace("data.table", quietly = TRUE)