+++ R/shape.r
@@ -1,9 +1,9 @@
 #' \code{shape}: prepare data for modeling with \code{dgirt} or \code{dgmrp}
 #'
-#' This function shapes various kinds of data for use in a dgirt model. Most
-#' arguments give the name or names of key variables in the data; they end in
-#' \code{_name} or \code{_names} and should be character vectors. Some others
-#' implement preprocessing and modeling choices.
+#' This function shapes data for use in a dgirt or dgmrp model. Most
+#' arguments give the name or names of key variables in the data. These
+#' arguments end in \code{_name} or \code{_names} and should be character
+#' vectors.
 #'
 #' @section Item Response Data:
 #' Individual-level data giving item responses is expected as argument
@@ -27,7 +27,7 @@
 #' The data given by \code{aggregate_data} must be in a long table of trial and
 #' success counts indexed by item, group, and time period. The variable names
 #' given by arguments \code{group_names}, \code{geo_name}, and\code{time_name}
-#' should exist in \code{aggregate_data}.  Three fixed variable names must also
+#' should exist in \code{aggregate_data}. Three fixed variable names must also
 #' appear in \code{aggregate_data}: \code{item} giving item identifiers,
 #' \code{n_grp} giving counts of item-response trials, and \code{s_grp} giving
 #' counts of item-response successes. These counts should be adjusted
@@ -138,9 +138,6 @@
 #' # check sparseness of data to be modeled
 #' get_item_n(shaped_responses, by = "year")
 #'
-#' @import data.table
-#' @seealso \code{\link{dgirtin-class}} \code{\link{dgirtfit-class}}
-#' @include require_namespace.r
 #' @export
 shape <- function(item_data,
                   item_names,
@@ -191,7 +188,11 @@ shape <- function(item_data,
                        constant_item = constant_item,
                        ...)
 
-  d_in <- dgirtIn$new(ctrl)
+  # setDT copies of *_data 
+  item_data <- set_copy_dt(item_data)
+  aggregate_data <- set_copy_dt(aggregate_data)
+  modifier_data <- set_copy_dt(modifier_data)
+  target_data <- set_copy_dt(target_data)
 
   # validate inputs #
   check_targets(target_data, ctrl)
@@ -203,11 +204,8 @@ shape <- function(item_data,
   item_data <- restrict_items(item_data, ctrl)
   ctrl@item_names <- intersect(ctrl@item_names, names(item_data))
   aggregate_data <- restrict_aggregates(aggregate_data, ctrl)
-  ctrl@aggregate_item_names <-
-    ctrl@aggregate_item_names[ctrl@aggregate_item_names %chin%
-                              aggregate_data$item]
-  d_in$time_observed <- get_observed(item_data, aggregate_data, ctrl@time_name)
-  d_in$geo_observed <- get_observed(item_data, aggregate_data, ctrl@geo_name)
+  ctrl@aggregate_item_names <- intersect(ctrl@aggregate_item_names,
+    aggregate_data$item)
 
   # rake survey weights #
   if (length(target_data)) {
@@ -215,11 +213,33 @@ shape <- function(item_data,
     ctrl@weight_name <- "raked_weight"
   }
 
+  d_in <- init_dgirt_in(item_data, aggregate_data, modifier_data, target_data,
+    ctrl)
+  d_in$call <- match.call()
+  d_in$pkg_version <- packageVersion("dgo")
+
+  # validate input to model #
+  check_dimensions(d_in)
+  check_values(d_in)
+  check_names(d_in)
+
+  d_in
+}
+
+set_copy_dt <- function(input_data) {
+  if (length(input_data)) {
+    return(data.table::setDT(data.table::copy(input_data)))
+  } else {
+    return(NULL)
+  }
+}
+
+init_dgirt_in <- function(item_data, aggregate_data, modifier_data, target_data,
+  ctrl) {
+  d_in <- dgirtIn$new(ctrl)
+
   # aggregate individual item response data to group level #
   item_data <- dichotomize(item_data, ctrl)
-  # this assignment should be redundant, but without it some variables created
-  # in dichotomize() weren't appearing in item_data
-
   d_in$group_grid <- make_group_grid(item_data, aggregate_data, ctrl)
   d_in$group_grid_t <- make_group_grid_t(d_in$group_grid, ctrl)
   d_in$group_counts <- make_group_counts(item_data, aggregate_data, ctrl)
@@ -235,23 +255,23 @@ shape <- function(item_data,
   d_in$s_vec <- setNames(d_in$group_counts$s_grp, d_in$group_counts$name)
 
   d_in$G <- nrow(unique(d_in$group_counts[, c(ctrl@geo_name, ctrl@group_names),
-                        with = FALSE]))
+      with = FALSE]))
   d_in$G_hier <- ifelse(!length(modifier_data), nlevels(gl(1L, d_in$G)),
-                        max(unlist(length(ctrl@modifier_names)), 1L))
+    max(unlist(length(ctrl@modifier_names)), 1L))
   d_in$T <- length(ctrl@time_filter)
   d_in$Q <- length(d_in$gt_items)
 
+  # not yet implemented
   d_in$WT <- array(1, dim = c(d_in$T, d_in$G_hier, d_in$G))
-
   d_in$l2_only <- matrix(0L, nrow = length(ctrl@time_filter), ncol = d_in$Q)
   d_in$NNl2 <- array(0L, dim = c(d_in$T, d_in$Q, d_in$G_hier))
   d_in$SSl2 <- d_in$NNl2
 
   d_in$XX <- make_design_matrix(d_in, ctrl)
-  d_in$ZZ <- shape_hierarchical_data(item_data, modifier_data, d_in, ctrl,
-                                     t1 = FALSE)
-  d_in$ZZ_prior <- shape_hierarchical_data(item_data, modifier_data, d_in, ctrl,
-                                           t1 = TRUE)
+  d_in$ZZ <- shape_hierarchical_data(modifier_data, ctrl@modifier_names,
+    d_in$group_grid_t, d_in$XX, ctrl)
+  d_in$ZZ_prior <- shape_hierarchical_data(modifier_data,
+    ctrl@t1_modifier_names, d_in$group_grid_t, d_in$XX, ctrl)
   d_in$hier_names <- dimnames(d_in$ZZ)[[2]]
 
   d_in$D <- ifelse(ctrl@constant_item, 1L, d_in$T)
@@ -259,9 +279,7 @@ shape <- function(item_data,
   d_in$P <- ncol(d_in$ZZ)
   d_in$S <- length(unique(d_in$group_grid[[ctrl@geo_name]])) - 1
   d_in$H <- dim(d_in$ZZ)[[3]]
-  #d_in$Hprior <- d_in$H
   d_in$Hprior <- dim(d_in$ZZ_prior)[[3]]
-
 
   # include subset data and other objects that may be useful later #
   d_in$item_data <- item_data
@@ -269,85 +287,7 @@ shape <- function(item_data,
   d_in$aggregate_data <- aggregate_data
   d_in$target_data <- target_data
   d_in$control <- ctrl
-  d_in$call <- match.call()
-  d_in$pkg_version <- packageVersion("dgo")
-
-  # validate input to model #
-  check_dimensions(d_in)
-  check_values(d_in)
-  check_names(d_in)
-
-  d_in
-}
-
-shape_hierarchical_data <- function(item_data, modifier_data, d_in, ctrl, t1) {
-
-  if (isTRUE(t1)) {
-    modifier_names <- ctrl@t1_modifier_names
-  } else {
-    modifier_names <- ctrl@modifier_names
-  }	
-  if (!length(modifier_names)) {
-    modifier_names <- NA
-  }
-
-  if (!length(modifier_data) | any(is.na(modifier_names))) {
-      zz.names <- list(ctrl@time_filter, dimnames(d_in$XX)[[2]], "")
-      zz <- array(data = 0, dim = lapply(zz.names, length), dimnames = zz.names)
-    } else {
-    # the array of hierarchical data ZZ should be T x P x H, where T is the
-    # number of time periods, P is the number of hierarchical parameters
-    # (including the geographic), and H is the number of predictors for
-    # geographic unit effects
-
-    hier_frame <- data.table::copy(modifier_data)
-
-    extra_colnames <- setdiff(names(hier_frame),
-                             c(ctrl@geo_name, ctrl@time_name, modifier_names))
-    if (length(extra_colnames)) {
-      hier_frame[, c(extra_colnames) := NULL]
-    }
-
-    # NOTE: We create param by renaming geo_name. Thus the requirement to model
-    # geographic predictors.
-    hier_frame[, c("param", ctrl@geo_name) := list(hier_frame[[ctrl@geo_name]],
-                                                   NULL)]
-    data.table::setkeyv(hier_frame, c("param", ctrl@time_name))
-    all(ctrl@time_filter %in% hier_frame$D_year)
-
-    # "param" is just the unique values of the geo var at the moment
-    modeled_param_names <- unique(hier_frame[, get("param")])
-
-    # unmodeled param levels will be those of groups
-    unmodeled_param_levels = unlist(lapply(d_in$unmod_par_names, function(x) {
-                                             paste0(x, unique(d_in$group_grid_t[[x]]))[-1]
-        }))
-    param_levels <- c(modeled_param_names, unmodeled_param_levels)
-
-    # make a zeroed table for unmodeled parameters by time period
-    unmodeled_frame <- expand.grid(c(list(unmodeled_param_levels,
-                                          ctrl@time_filter), rep(list(0L),
-                                          length(modifier_names))))
-    unmodeled_frame <- setNames(unmodeled_frame, c("param", ctrl@time_name,
-                                                   modifier_names))
-    data.table::setDT(unmodeled_frame, key = c("param", ctrl@time_name))
-
-    hier_frame <- rbind(hier_frame, unmodeled_frame)
-
-    zz <- sapply(modifier_names, function(x) {
-                   matrix(hier_frame[[x]],
-                          # We have T rows, so filling by column is correct SO
-                          # LONG AS time varies fastest then param in
-                          # hier_frame
-                          nrow = length(unique(hier_frame[[ctrl@time_name]])),
-                          ncol = length(unique(hier_frame$param)),
-                          dimnames = list(unique(hier_frame[[ctrl@time_name]]),
-                                          unique(hier_frame$param)))
-                        }, simplify = 'array')
-    # omit first geo parameter
-    zz <- zz[, -1, , drop = FALSE]
-  }
-  zz
+  return(d_in)
 }
 
 make_design_matrix <- function(d_in, ctrl) {
@@ -368,8 +308,6 @@ make_design_matrix <- function(d_in, ctrl) {
                                  colnames(design_matrix))
 
   design_matrix <- subset(design_matrix, select = -1)
-
-  # TODO: move to S4 validate
   invalid_values <- setdiff(as.vector(design_matrix), c(0, 1))
   if (length(invalid_values)) {
     stop("design matrix values should be in (0, 1); found ",