+++ tests/testthat.r
@@ -1,4 +1,0 @@
-library(testthat)
-library(dgo)
-
-test_check("dgo")