+++ NEWS.md
@@ -1,3 +1,16 @@
+# 0.2.9
+
+* Switch from compiling Stan models at install time to compiling them at
+  runtime, avoiding an Rcpp module issue.
+* Add `model` argument to `dgirt()` and `dgmrp()` taking for reuse a previously
+  compiled Stan model, as found in the `@stanmodel` slot of a `dgirt_fit`- or
+  `dgmrp_fit`-class object.
+* The `version` argument to `dgirt()` and `dgmrp()` can be used to specify
+  arbitrary `.stan` files on the disk in addition to those included with the
+  package.
+* Argument `by` to `get_n()` and `get_item_n()` methods properly accepts a
+  vector of variable names when combined with `aggregate` arguments.
+
 # 0.2.8
 
 * Improve Stan models for shorter run times